mcp/requirements.txt
mcp>=1.1.0
httpx>=0.27
python-dotenv>=1.0
